added the basic templates then edited to suit list
added to basic templated for the CRUD functionality, the templates came from diary app thats why almost all the app looks the same. i replaced variable names and remove stuffs that list app dont need. in the future i will put more efforts on the appearance of the apps but for now i just want the core functionality to be functional, i need to use immediately