managed to fix create view for list entry
to create a list entry you need a list object because its pk is a list obj. The pre existing create view are all dependent on  the user id. So I need to find a way to get that specific list object, my first though is to use function based view just for the create view. But when a revisited my old code, it looks tedious to implement. Luckily through searching I found a way to implement it on createview so YATTTA!!!!