implemented the remaining CRUD functionality
just implemented the remaining CRUD functionality, hopefully it works. I made significant changes to facilitate the locations of list obj. The templates and urls are not yet done, so i wont know for now. The good thing is there is no error on the console so hoping for the best